sec-api-backend/pyproject.toml
[tool.poetry]
name = "sec-api-backend"
version = "0.1.0"
description = "FastAPI backend for SEC Drug Asset Extraction"
authors = ["sais07 <sai@cbase.ai>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.0,<3.10"
fastapi = "0.104.1"
uvicorn = "0.23.2"
python-dotenv = "1.0.0"
boto3 = "1.34.69"
botocore = "1.34.69"
s3transfer = "0.10.2"
sec-api = "1.0.12"
pandas = "2.1.4"
langchain = "0.2.5"
langchain-huggingface = "0.0.3"
langchain-openai = "0.1.14"
langchain-community = "0.2.5"
langchain-text-splitters = "0.2.1"
langchain-core = "0.2.8"
langsmith = "0.1.81"
beautifulsoup4 = "4.12.3"
python-multipart = "0.0.6"
tabulate = "0.9.0"
pydantic = "1.10.13"
pydantic-core = "2.18.1"
faiss-cpu = "1.8.0"
requests = "2.32.3"
setuptools = "75.8.0"
protobuf = "4.25.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.28.1"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"